Created reusable function for trpc response meta data
