+++ packages/trpc/react/trpc.ts
@@ -41,6 +41,13 @@ const ENDPOINTS = [
 export type Endpoint = (typeof ENDPOINTS)[number];
 
 const resolveEndpoint = (links) => {
+  // TODO: Fix this error
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  // TODO: Update our trpc routes so they aren't muddy.
+  // This function parses paths like the following are maps them
+  // to the correct API endpoints.
+  // - viewer.me - 2 segment paths like this are for logged in requests
+  // - viewer.public.i18n - 3 segments paths can be public or authed
   return (ctx) => {
     const parts = ctx.op.path.split(".");
     let endpoint = '';
@@ -52,6 +59,7 @@ const resolveEndpoint = (links) => {
       endpoint = parts[1];
       path = parts.splice(2, parts.length - 2).join('.');
     }
+
     return links[endpoint]({ ...ctx, op: { ...ctx.op, path } });
   };
 }
